Combined overlapping hipRtcTypeNames
