+++ dnn-providers/integration-tests/gpu_ref/include/hipdnn_gpu_ref/detail/HipRtcTypeName.hpp
@@ -1,0 +1,59 @@
+// Copyright © Advanced Micro Devices, Inc., or its affiliates.
+// SPDX-License-Identifier:  MIT
+
+#pragma once
+
+#include <hipdnn_data_sdk/types/Bfloat16.hpp>
+#include <hipdnn_data_sdk/types/Half.hpp>
+
+#include <cstdint>
+
+namespace hipdnn_gpu_ref::detail
+{
+
+template <typename T>
+struct HipRtcTypeName;
+
+template <>
+struct HipRtcTypeName<float>
+{
+    static constexpr const char* VALUE = "float";
+};
+
+template <>
+struct HipRtcTypeName<hipdnn_data_sdk::types::half>
+{
+    static constexpr const char* VALUE = "_Float16";
+};
+
+template <>
+struct HipRtcTypeName<hipdnn_data_sdk::types::bfloat16>
+{
+    static constexpr const char* VALUE = "__bf16";
+};
+
+template <>
+struct HipRtcTypeName<double>
+{
+    static constexpr const char* VALUE = "double";
+};
+
+template <>
+struct HipRtcTypeName<int8_t>
+{
+    static constexpr const char* VALUE = "signed char";
+};
+
+template <>
+struct HipRtcTypeName<uint8_t>
+{
+    static constexpr const char* VALUE = "unsigned char";
+};
+
+template <>
+struct HipRtcTypeName<int32_t>
+{
+    static constexpr const char* VALUE = "int";
+};
+
+} // namespace hipdnn_gpu_ref::detail
